Porting 32-bit code to 64-bit
Compiler flags:
 -Werror=pointer-to-int-cast
 -Werror=int-to-pointer-cast
 -Werror=shorten-64-to-32

See also: https://developer.android.com/distribute/best-practices/develop/64-bit

+++ include/tuner_e4k.h
@@ -180,9 +180,9 @@ enum e4k_if_filter {
 	E4K_IF_FILTER_RC
 };
 struct e4k_pll_params {
-	uint32_t fosc;
+	uint64_t fosc;
 	uint32_t intended_flo;
-	uint32_t flo;
+	uint64_t flo;
 	uint16_t x;
 	uint8_t z;
 	uint8_t r;
@@ -204,8 +204,8 @@ int e4k_if_gain_set(struct e4k_state *e4k, uint8_t stage, int8_t value);
 int e4k_mixer_gain_set(struct e4k_state *e4k, int8_t value);
 int e4k_commonmode_set(struct e4k_state *e4k, int8_t value);
 int e4k_tune_freq(struct e4k_state *e4k, uint32_t freq);
-int e4k_tune_params(struct e4k_state *e4k, struct e4k_pll_params *p);
-uint32_t e4k_compute_pll_params(struct e4k_pll_params *oscp, uint32_t fosc, uint32_t intended_flo);
+uint64_t e4k_tune_params(struct e4k_state *e4k, struct e4k_pll_params *p);
+uint64_t e4k_compute_pll_params(struct e4k_pll_params *oscp, uint64_t fosc, uint32_t intended_flo);
 int e4k_if_filter_bw_get(struct e4k_state *e4k, enum e4k_if_filter filter);
 int e4k_if_filter_bw_set(struct e4k_state *e4k, enum e4k_if_filter filter,
 		         uint32_t bandwidth);
